fix hightlighting issue when working with ack options
I came across this bug today: set g:ackhighlight = 1 and do
 :Ack -i something

This will result in every "w" being highlighted in your current window
and in the quick fix list. I "fixed" this by trying to strip out any
options before the actual search string.

The code is a bad hack and if you happen to also specifiy directories
for ack to search in, they will still be highlighted. But the code kind
of works and I didn't want to simply open an issue.