Centralize state concerning list window type
Pulls script-global state into one place, less stringing along flags for
which list window type is in use.

Doesn't help to reduce the use of script-global state, but this is
starting to look like a refactoring step toward creating a search object
for each `:Ack*` invocation with encapsulated state, which could be a
nice way to go.