Add an explicit --nopager to default ack options
Otherwise if a user has something like `--pager=less` as a default in
their `~/.ackrc`, Ack may think that it's running interactively if
executed in a separate shell e.g. when `g:ack_use_dispatch` is enabled.
This will result in broken behavior loading results in Vim.

See #182 for some background.