Do everything using setloclist/setqflist
Use setloclist/setqflist not only to change the title, but also to
append/replace entries in the quickfix/loclist.

There was an issue earlier where entries in the quickfix/loclist windows
sometimes ended up not being properly formatted (somehow errorformat was
not properly picked up); this new solution should hopefully fix that.