Map uniquely
If a previous mapping using the provided key sequence is defined, the
map will fail. silent! prevents this from being a problem (stopping
execution or displaying annoying errors).

The long-term effect is to avoid overwriting user mappings in the qf
window.

A second solution would be, for each `{map} {lhs} {rhs}` command, to use

    try
      {map} <unique> {lhs} {rhs}
    catch /E227/
    endtry

which accomplishes the same effect less concisely.