Fix support for using ripgrep with vim-dispatch
Same as ag, [ripgrep](https://github.com/BurntSushi/ripgrep) supports a
`--vimgrep` flag so that it can be used as `g:ackprog`.

However, when used via `vim-dispatch`, stdin is not closed which ripgrep
interprets as a pattern to be supplied and hence fails to execute
anything.

To disable this behavior, we need to explicitly supply a root-directory
to start searching in.