"Fix" outputting to terminal every search!
I wouldn't say it fixes #18 and #52, but it builds on them. It avoids
the compatibility woes, but only using the shell trick if the shell is
bash or zsh. Otherwise it does the t_te/i swapping.

Both are not very neat, and make this function messier, but making 3
separate functions seemed a lot more busy work.